refactor greetserver
